changed versioning for uploading new plugin
