+++ gradle.properties
@@ -4,7 +4,7 @@ pluginGroup = com.github.moreaframework.intellijplugin
 pluginName = Morea
 pluginRepositoryUrl = https://github.com/morea-framework/plugin-intellij
 # SemVer format -> https://semver.org
-pluginVersion = 0.0.3
+pluginVersion = 1.0.0
 
 # Supported build number ranges and IntelliJ Platform versions -> https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
 pluginSinceBuild = 213
