+++ src/main/java/com/moreaframework/NewOutcomesAction.java
@@ -1,4 +1,4 @@
-package com.MoreaFramework;
+package com.moreaframework;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
@@ -11,12 +11,12 @@ import com.intellij.openapi.vfs.VirtualFile;
 import java.io.IOException;
 import java.io.OutputStream;
 
-public class NewAssessmentsAction extends AnAction {
+public class NewOutcomesAction extends AnAction {
 
     @Override
     public void actionPerformed(AnActionEvent e) {
         MoreaUtils morea = new MoreaUtils();
-        String input = Messages.showInputDialog("Enter Assessments Page Name:", "New Assessments Page", Messages.getQuestionIcon(), "assessment-example", new InputValidator() {
+        String input = Messages.showInputDialog("Enter Outcomes Page Name:", "New Outcomes Page", Messages.getQuestionIcon(), "outcome-example", new InputValidator() {
             @Override
             public boolean checkInput(String inputString) {
                 return !inputString.isEmpty();
@@ -31,7 +31,7 @@ public class NewAssessmentsAction extends AnAction {
         // Check if name to create is not null
         if (input != null) {
             // Check input name matches MOREA naming convention
-            input = morea.toMoreaName(input, "assessment");
+            input = morea.toMoreaName(input, "outcome");
             VirtualFile directory = morea.checkDupes(e, input);
 
             String finalInput = input;
@@ -43,14 +43,14 @@ public class NewAssessmentsAction extends AnAction {
                                 "title: \"CHANGE ME\"\n" +
                                 "published: true\n" +
                                 "morea_id: " + finalInput + "\n" +
-                                "morea_type: assessment\n" +
+                                "morea_type: outcome\n" +
                                 "morea_summary: \"CHANGE ME\"\n" +
                                 "morea_sort_order: \n" +
                                 "morea_start_date: \n" +
                                 "morea_labels: \n" +
                                 "---\n\n" +
                                 "## \"CHANGE ME\"\n\n" +
-                                "This is a sample content for the newly creating assessment.md file";
+                                "This is a sample content for the newly creating outcome.md file";
                         outputStream.write(content.getBytes());
                     } catch (IOException ex) {
                         ex.printStackTrace();
