+++ src/main/java/com/moreaframework/NewPageAction.java
@@ -1,4 +1,4 @@
-package com.MoreaFramework;
+package com.moreaframework;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
