+++ src/main/java/com/moreaframework/settings/AppSettingsState.java
@@ -1,4 +1,4 @@
-package com.MoreaFramework.settings;
+package com.moreaframework.settings;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.components.PersistentStateComponent;
